// src/tickets/interfaces/ticket-responde.interface.ts
export class TicketResponse {
  t2_id?: number;
  t2_fecha?: string;
  t2_codigo?: string;
  t2_servicio?: string;
  t2_tipo?: string | null;
  t2_unq?: string;
  t1_alumnos_t1_id?: number;
  t3_periodos_t3_id?: number;
  t1_id?: number;
  t1_dni?: string;
  t1_codigo?: string;
  t1_nombres?: string;
  t1_escuela?: string;
  t1_unq?: string;
  t1_estado?: number;
  code: number;
}